fixing issues related to contest life cycle

+++ app/backends/webapp/src/main/scala/com/shocktrade/webapp/vm/opcodes/CreditWallet.scala
@@ -11,7 +11,7 @@ import scala.concurrent.{ExecutionContext, Future}
  */
 class CreditWallet(portfolioID: String, amount: Double) extends OpCode {
 
-  override def invoke()(implicit ctx: VirtualMachineContext, ec: ExecutionContext): Future[Int] = {
+  override def invoke()(implicit ctx: VirtualMachineContext, ec: ExecutionContext): Future[Double] = {
     ctx.creditWallet(portfolioID, amount)
   }
